mid refactor but going to remove hyper data
