+++ autoarray/inversion/inversion/factory.py
@@ -10,8 +10,8 @@ from autoarray.structures.visibilities import VisibilitiesNoiseMap
 from autoarray.operators.convolver import Convolver
 from autoarray.operators.transformer import TransformerDFT
 from autoarray.operators.transformer import TransformerNUFFT
-from autoarray.inversion.linear_obj import LinearObj
-from autoarray.inversion.linear_obj import LinearObjFunc
+from autoarray.inversion.linear_obj.func_list import LinearObj
+from autoarray.inversion.linear_obj.func_list import LinearObjFuncList
 from autoarray.inversion.linear_eqn.imaging.w_tilde import LEqImagingWTilde
 from autoarray.inversion.linear_eqn.imaging.mapping import LEqImagingMapping
 from autoarray.inversion.inversion.matrices import InversionMatrices
@@ -82,7 +82,7 @@ def inversion_imaging_unpacked_from(
     profiling_dict: Optional[Dict] = None,
 ):
 
-    if any(isinstance(linear_obj, LinearObjFunc) for linear_obj in linear_obj_list):
+    if any(isinstance(linear_obj, LinearObjFuncList) for linear_obj in linear_obj_list):
         use_w_tilde = False
     elif preloads.use_w_tilde is not None:
         use_w_tilde = preloads.use_w_tilde
@@ -144,7 +144,7 @@ def inversion_interferometer_unpacked_from(
     except ImportError:
         settings.use_w_tilde = False
 
-    if any(isinstance(linear_obj, LinearObjFunc) for linear_obj in linear_obj_list):
+    if any(isinstance(linear_obj, LinearObjFuncList) for linear_obj in linear_obj_list):
         use_w_tilde = False
     else:
         use_w_tilde = settings.use_w_tilde
