+++ autoarray/inversion/linear_eqn/interferometer/w_tilde.py
@@ -7,7 +7,7 @@ from autoarray.inversion.linear_eqn.interferometer.abstract import (
     AbstractLEqInterferometer,
 )
 from autoarray.dataset.interferometer import WTildeInterferometer
-from autoarray.inversion.linear_obj import LinearObj
+from autoarray.inversion.linear_obj.func_list import LinearObj
 from autoarray.inversion.inversion.settings import SettingsInversion
 from autoarray.preloads import Preloads
 from autoarray.operators.transformer import TransformerNUFFT
