+++ to_do_list
@@ -22,7 +22,7 @@ __Linear obj Func + Mapper fast leq__
 - Create MapperLeq object, which creates curvature / regularization / curvature reg matrix for each mapper (could
 be Mapping or WTilde object).
 
-- Extend LinearObjFunc object to have curvature / regularization / curvature reg matrix, where it always uses
+- Extend LinearObjFuncList object to have curvature / regularization / curvature reg matrix, where it always uses
 Mapping calc and regularization matrix is all zeros.
 
 - Create LeqOffDiag objects, which take two of the above and compute curvature / regularization / curvature reg matrix
@@ -47,7 +47,7 @@ above objects, instead of constructing a large `blurred_mapping_matrix` (albeit 
 - Could aim for LEqImaging / LEqInterferometer to be reduced to just LEq,
 
 - For efficiency, you prob need to preallocate curvature_reg / curvature_reg_mapper (or perhaps even all
-matrices of every individual MapperLEq / LinearObjFunc / LeqOffDiag object?) and copy in memory.
+matrices of every individual MapperLEq / LinearObjFuncList / LeqOffDiag object?) and copy in memory.
 
 
 __Probably Terrible Idea Notes__
